chore: added feature flag, applied PR suggestions
