+++ apps/browser-extension-wallet/src/views/browser-view/features/sing-message/SignMessageDrawer.tsx
@@ -1,31 +1,37 @@
 import React from 'react';
 import { useSignMessageState } from './useSignMessageState';
-import { SignMessage } from './SignMessage';
+import { useDrawerConfiguration } from './useDrawerConfiguration';
+import { SignMessageForm } from './SignMessageForm';
+import { SignMessageResult } from './SignMessageResult';
+import styles from './SignMessageDrawer.module.scss';
 
 export const SignMessageDrawer: React.FC = () => {
   const {
     usedAddresses,
     isSigningInProgress,
-    signature,
+    signatureObject,
     error,
     hardwareWalletError,
     isHardwareWallet,
-    performSigning,
-    closeDrawer
+    performSigning
   } = useSignMessageState();
 
+  useDrawerConfiguration();
+
   return (
-    <>
-      <SignMessage
-        addresses={usedAddresses}
-        onClose={closeDrawer}
-        onSign={performSigning}
-        isSigningInProgress={isSigningInProgress}
-        signedMessage={signature}
-        error={error}
-        hardwareWalletError={hardwareWalletError}
-        isHardwareWallet={isHardwareWallet}
-      />
-    </>
+    <div data-testid="sign-message" className={styles.container}>
+      {signatureObject?.signature ? (
+        <SignMessageResult signature={signatureObject.signature} />
+      ) : (
+        <SignMessageForm
+          usedAddresses={usedAddresses}
+          isSigningInProgress={isSigningInProgress}
+          error={error}
+          hardwareWalletError={hardwareWalletError}
+          isHardwareWallet={isHardwareWallet}
+          performSigning={performSigning}
+        />
+      )}
+    </div>
   );
 };
