--- apps/browser-extension-wallet/src/views/browser-view/features/sing-message/SignMessagePassword.module.scss
@@ -1,44 +1,0 @@
-@import '../../../../../../../packages/common/src/ui/styles/theme.scss';
-@import '../../../../../../../packages/common/src/ui/styles/abstracts/typography';
-.passwordContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 24px;
-}
-
-.title {
-  font-size: 20px;
-  font-weight: bold;
-  margin-bottom: 16px;
-}
-
-.subtitle {
-  font-size: 14px;
-  color: #666;
-  margin-bottom: 24px;
-  text-align: center;
-}
-
-.passwordWrapper {
-  width: 100%;
-  margin-bottom: 24px;
-}
-
-.buttonContainer {
-  display: flex;
-  flex-direction: column;
-  gap: size_unit(2);
-  margin-top: size_unit(3);
-}
-
-.confirmButton,
-.cancelButton {
-  padding: 12px;
-  border-radius: 4px;
-  font-size: 16px;
-  font-weight: bold;
-  cursor: pointer;
-  transition: background-color 0.3s;
-  margin-bottom: 12px;
-}
