+++ apps/browser-extension-wallet/src/views/browser-view/features/sing-message/useSignMessageState.tsx
@@ -1,31 +1,28 @@
 import { useState, useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useWalletStore } from '@stores';
-import { PostHogAction, useObservable } from '@lace/common';
+import { useObservable } from '@lace/common';
 import { withSignDataConfirmation } from '@lib/wallet-api-ui';
 import { Cip30DataSignature } from '@cardano-sdk/dapp-connector';
 import { Wallet } from '@lace/cardano';
 import { HexBlob } from '@cardano-sdk/util';
-import { useDrawer } from '@views/browser/stores';
-import { useAnalyticsContext } from '@providers';
+// import { useAnalyticsContext } from '@providers';
 
 interface SignMessageState {
   usedAddresses: { address: string; id: number }[];
   isSigningInProgress: boolean;
-  signature: Cip30DataSignature | undefined;
+  signatureObject: Cip30DataSignature | undefined;
   error: string;
   hardwareWalletError: string;
   isHardwareWallet: boolean;
   performSigning: (address: string, message: string, password: string) => void;
-  closeDrawer: () => void;
 }
 
 export const useSignMessageState = (): SignMessageState => {
   const { t } = useTranslation();
-  const analytics = useAnalyticsContext();
+  // const analytics = useAnalyticsContext();
 
   const { inMemoryWallet, isHardwareWallet } = useWalletStore();
-  const [, setDrawerConfig] = useDrawer();
   const [isSigningInProgress, setIsSigningInProgress] = useState(false);
   const [signature, setSignature] = useState<Cip30DataSignature>();
   const [error, setError] = useState<string>('');
@@ -74,10 +71,10 @@ export const useSignMessageState = (): SignMessageState => {
     [inMemoryWallet, isHardwareWallet, t]
   );
 
-  const closeDrawer = useCallback(() => {
-    analytics.sendEventToPostHog(PostHogAction.SignMessageCloseDrawer);
-    setDrawerConfig();
-  }, [setDrawerConfig, analytics]);
+  // const closeDrawer = useCallback(() => {
+  //   analytics.sendEventToPostHog(PostHogAction.SignMessageCloseDrawer);
+  //   setDrawerConfig();
+  // }, [setDrawerConfig, analytics]);
 
   const usedAddresses =
     addresses?.map((address, index) => ({
@@ -85,14 +82,46 @@ export const useSignMessageState = (): SignMessageState => {
       id: index
     })) || [];
 
+  // return useCallback(() => {
+  //   setDrawerConfig({
+  //     renderHeader: () => <HeaderNavigation isPopupView={isPopupView} flow={content} />,
+  //     renderTitle: () => <HeaderTitle popup={isPopupView} />,
+  //     ...(shouldRenderFooter
+  //       ? {
+  //           renderFooter: () => (
+  //             <Footer
+  //               key={currentSection}
+  //               isPopupView={isPopupView}
+  //               onHandleChangeConfirm={(action) => {
+  //                 action === 'DELETE' && setAddressValue(row, '');
+  //                 setSection({ currentSection: Sections.FORM, nextSection: Sections.SUMMARY });
+  //               }}
+  //             />
+  //           )
+  //         }
+  //       : {}),
+  //     ...config
+  //   });
+  // }, [
+  //   config,
+  //   content,
+  //   currentSection,
+  //   isPopupView,
+  //   onClose,
+  //   row,
+  //   setAddressValue,
+  //   setDrawerConfig,
+  //   setSection,
+  //   shouldRenderFooter
+  // ]);
+
   return {
     usedAddresses,
     isSigningInProgress,
-    signature,
+    signatureObject: signature,
     error,
     hardwareWalletError,
     isHardwareWallet,
-    performSigning,
-    closeDrawer
+    performSigning
   };
 };
