+++ packages/translation/src/lib/translations/core/en.json
@@ -202,26 +202,21 @@
   "core.signMessage.selectAddress": "Select an address to use",
   "core.signMessage.messageLabel": "Message to sign",
   "core.signMessage.messagePlaceholder": "Paste or type your message here",
-  "core.signMessage.submit": "Submit",
   "core.signMessage.incorrectPassword": "Incorrect password. Please try again.",
   "core.signMessage.signingInProgress": "Signing in progress",
   "core.signMessage.signatureLabel": "Signature",
-  "core.signMessage.enterPasswordTitle": "Enter Password",
-  "core.signMessage.enterPasswordSubtitle": "Please enter your password to sign the message",
   "core.signMessage.passwordLabel": "Password",
   "core.signMessage.walletNotInitialized": "Wallet not initialized",
   "core.signMessage.signingFailed": "Signing failed",
-  "core.signMessage.emptyAddressOrMessage": "Address or message is empty",
-  "core.signMessage.hardwareWalletPrompt": "Hardware Wallet Confirmation",
   "core.signMessage.signWithHardwareWallet": "Sign with Hardware Wallet",
   "core.signMessage.pleaseConfirmOnDevice": "Please confirm the signing operation on your hardware wallet device.",
   "core.signMessage.copyToClipboard": "Copy to Message",
   "core.signMessage.successTitle": "All done!",
-  "core.signMessage.successDescription": "This is your encrypted signed message",
+  "core.signMessage.successDescription": "This the signature of your signed message",
   "core.signMessage.nextButton": "Next",
   "core.signMessage.closeButton": "Close",
-  "core.signMessage.signButton": "Sign Message",
-  "core.signMessage.enterPassword": "Enter your wallet password",
+  "core.signMessage.signButton": "Sign message",
+  "core.signMessage.signWithHardwareWalletButton": "Sign message with hardware wallet",
   "core.signMessage.hardwareWalletNotConnected": "Make sure your hardware wallet is connected and unlocked.",
   "core.Burn.title": "Confirm transaction",
   "core.coinInputSelection.assetSelection": "Select tokens or NFTs",
